LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-placeholder-identity-expected.txt


FAIL Placeholder CSSStyleSheet is replaced by a different object after fetch. assert_equals: Before fetch: expected 1 placeholder(s). expected 1 but got 0
